Create a file for the tip of each branch.
Create a file of the form project-branch.tar.gz so that there is a
url that can be referenced for each projects that represents a proper
source sdist for the tip of the branch. That way, if someone (such as
ceilometer) needs to depend on tip of nova, they can put in:

http://nova.openstack.org/tarballs/nova-master.tar.gz

In their pip-requires file instead of a github zipball link (which does not
have all of the files in the manifest, since it's not a proper sdist.

Change-Id: I0228381894417a15655e7b684c5a65f98a2fc7ba