Put jenkins:java in the unlimited cgroup.
Previously attempted to put jenkins:jenkins in the unlimited cgroup
(so that the jenkins slave administrative stuff didn't count against
the tests' limits), but the jenkins slave process runs as `java -jar
slave.jar` instead. This means jenkins:java needs to be in its own
cgroup to get the same effect.

Change-Id: Ib3f8f765169a0b85f21ac59ae0177865210f8649