Trim down some un-needed repetition.
The close_pull was an extra argument that was just verbose and never
used in any way other than true.
Adding a yml file to the project but not listing it in the .pp file
only really ever caused bugs.
Moved the full openstack project list to a variable for re-use.

Change-Id: I211b5cf54d5a84c6136a4b04ea819c688cc1e2b6