Add script to configure gerrit project ACLs.
Add a script to the Gerrit module to configure Gerrit project ACLs.
The script expects the project yaml to exist and the environment
variable pointing at it to be valid. The script takes two arguments.
First the directory in which to make the change in (projects will be
cloned into this directory). The second is the base url from which to
clone from. It should end in a trailing '/'.

Change-Id: I75c545b00c1daa5829a2dab5825a9262bd3b2e1f