Change the name of the pypi tarball job.
The pypi_jobs template defines a -tarball job which is different
than the -tarball job in python_jobs template.  That's confusing,
and not supported by the version of job-builder in development.

This changes the name of that job, and updates the zuul layout.

This alse enables the reddwarf jobs, which are configured in zuul.

Also, there's a hackish change to test.sh and the test output
to write out xml files in a way that's compatible with the
version of job-builder in development, so that they may be
directly compared.  The new version will not implement that
in quite such a hackish way.

Change-Id: Ice3f03b41e44e2979942f516febd059f4ab4cefa