New job template for translation management.
Add a new job template for translation management. Adds a new job that
will pull the latest translations from transifex, updated the .pot
file, update all of the .po files using the new .pot file, add all of
the changed files to git then commit and review, and finally push the
updated .pot and .po files to transifex.

Apply the template to nova.

Change-Id: I5242f81bd6ff13d1ed8610ef9e3b39fd3dbd7301