New job template for translation management.
Add a new job template for translation management. Add two jobs for
translation management. The first will will update transifex with the
new .pot and .po files after every merge to a project. The second will
commit the new .pot and .po files and submit for review once a day.

Apply the template to nova.

Change-Id: I5242f81bd6ff13d1ed8610ef9e3b39fd3dbd7301