New job template for translation management.
Add a new job template for translation management. It adds two new
jobs. The first updates transifex with new .pot files for projects.
The second pulls new translation results from transifex and adds them
to the tree. In both cases the git repo is updated and the changes are
pushed to gerrit for review.

Apply the template to nova.

Change-Id: I5242f81bd6ff13d1ed8610ef9e3b39fd3dbd7301