Ensure gerritbot is present and not latest.
Pip doesn't appear to be able to do a clean upgrade of gerritbot on
oneiric. It should remove the old versions of gerritbot after installing
the new versions but it does not. This confuses Puppet into thinking
that it need to reinstall gerritbot 0.1.2 on each puppet run which
restarts the IRC bot. To prevent this behavior ensure the package is
present and not latest.

Change-Id: I7729d126a7811b7e08541ed65ca48f4dc302ff6b