Adds a markdown to docbook script.
Patch set 3 makes the jobname the filename and adds a second job for
building the programming api mdown book.

Patch set 4 corrects the job name to match the file name.

Patch set 5 renames the .sh.save file and adds it to the job builders.

Patch set 6 addresses final comments - appending to the gerrit-doc.properties
file and removes tabs.

Patch set 7 addresses comments. Would like to continue work on markdown to
docbook and delay work on the properties script if possible.

Change-Id: I46c07cd11e96e8fc463a377519926a43981e4b3d