Add script to log jenkins sudo attempts.
By running this script before and after test runs, we can see if
there have been any sudo attempts by jenkins.  If there have been,
we can fail the test run.  The script outputs any new jenkins sudo
attempts found, so it will show up in the console log.

The script needs to be run as root, so a sudoers file is added
to allow jenkins to run it.

Change-Id: I4a4373d085ff7717d022880b3ab6110100371c4c