Use console-log scp macro to copy build logs.
Previoulsy we had spelled out that the build logs should be copied when
there were other files to copy in order to avoid multiple SCP
connections. Now, in order to copy the entire console log, we need
multiple SCP connections anyways so go back to using the console-log
macros for simplicity and consistency.

Change-Id: I6e88d08f7eec945b907353ed90ac0d482b7a9c9a