Propose translations based on master.
Previously if an existing transifex translation change was present the
translation proposal script would make changes against that change.
git-review would break and this ends up not actually proposing changes
that we want. To correct this get the change id of the existing change
and use that in the commit message for changes based off of master. This
gives us translations based on the latest code with the benefits of less
noise in gerrit.

Change-Id: I88bf13193c6541806388dd610c68e0a95205973e