set +e around read in propose_translation_update.
The bash read builtin returns a non zero value when it reads EOF.
Because the script uses a heredoc read will always read an EOF and
always return a non zero value. This non zero value and set -e causes
the script to exit. Temporarily set +e to get around this.

Change-Id: I2fc2f57ec492ce05089b15bd5b721d27ad4b2be3