Push src and tx files to transifex separately.
Push the source (.pot) file independently of the translation (.po) files
so that we can disable set -e when pushing the .po files.

Previously pushing the .po files would fail the entire test if one of
the pushed files contained no translations. This would also prevent
other translation files from being pushed. Now push the .pot file
separately so that a failure pushing that file still fails the test, but
errors pushing the translation files can be ignored. Also add the --skip
option to the translation push which will allow all valid translations
to be pushed even if some fail.

Change-Id: I8958c29113ce5d21a0b281ee13abe94d477fb04c