Publish python-*client docs only on releases.
Modify the run-docs script to upload the latest tagged release docs to
the root dir of a projects developer docs and to a dir for the tagged
release. With this in place and the python-*client-docs jobs running
only in the publish queue we will see documents at
docs.openstack.org/developer that follow the latest releases for the
client projects.

Finally, re add the python-*client docs jobs as they were the projects
that wanted this behavior in the first place.

Fixes bug #1040251

Change-Id: Ia3cf07d860b10214ed54fcf9748330866cb47f5d