Make it optional to always publish docs.
Previously the root dir for the developer docs was always published to
by the project-docs publisher job. Modify the run-docs script to make
that optional. For this to make sense also upload the latest tagged
releases to root dir of the developer docs.

Finally, readd the python-*client docs jobs as they were the projects
that wanted this behavior in the first place.

Fixes bug #1040251

Change-Id: Ia3cf07d860b10214ed54fcf9748330866cb47f5d