Gerrit Whitespace Change Detection
Consider whitespace changes significant enough not to reapply code
reviews, but still comment if that's all which changed between
patchsets. This addresses bug 1057506.

* modules/openstack_project/files/gerrit/scripts/trivial_rebase.py
(GetPatchId): Add a flag called ignore_whitespace, but defaulting to
True so as to preserve default behavior of the module. Add conditional
behavior to replace all spaces and tabs with percent signs before
calculating the patch-id hash.
(Main): Add a --whitespace command-line option to turn on whitespace
change checking. If enabled and if normal GetPatchId calls return a
match, re-run with ignore_whitespace set to False and apply a comment to
the new patchset in Gerrit if the result is non-matching.

* modules/openstack_project/templates/gerrit_patchset-created.erb: Add
--whitespace to the trivial_rebase.py invocation, enabling whitespace
checking.

Change-Id: I89c479614a637717cf515a5d3f6d03b5f7830581