Update run-docs to handle old python-*client tags
Some of the python-*client repos have year.release tags which are not
used anymore. These tags are always greater than the current versioning
scheme and thus break the logic in run-docs.sh of taking the largest tag
number as the latest tag. Ignore any tag beginning with '2012.' to when
running this script on the python-*client projects.

Also, since I was editing the file I ran s/GERRIT/ZUUL/ to update the
environment variables to the ones we want to be using in the future.

Change-Id: I6f7a5aeddd4a8fd69f25382682fceebf94bea629