Testing git symbolic heads on review-dev.
* modules/openstack_project/files/gerrit/review_dev_symbolic_heads.yaml:
A sample YAML file describing several arbitrary symbolic references for
branches in two existing test projects on review-dev.openstack.org.

* modules/openstack_project/files/gerrit/scripts/symbolic_heads.py: A
script which, when given a YAML file describing git symbolic references
for alternate heads in a list of repositories, will apply the necessary
actions to make the symbolic refs on disk correspond. This script can be
run repeatedly, and will only make adjustments when the on-disk
repositories do not match the supplied description file.

* modules/openstack_project/manifests/review_dev.pp: Adds the
python-git and python-yaml packages required by symbolic_heads.py,
installs that script and the symbolic_heads.yaml description file, and
executes the former with the latter as an argument.

Change-Id: Ib46b5a6524e69c7d4c1ebb443159296844dbc96f