Add a named tag manager to review-dev.
A mechanism for applying a tag->branch mapping in a per-project YAML
file stored in a particular branch or meta-branch, triggered as a Gerrit
ref-updated hook module. This should allow release managers to maintain
symbolic names which transition from one branch to another during each
project's lifecycle.

* modules/openstack_project/files/gerrit/scripts/update_nametags.py: A
script which reads a YAML description file found in a particular git
branch or meta-branch and then updates lightweight git tags if the
current change was merged into a branch mentioned therein. It will also
update tags for all branches mentioned if the meta-branch itself is what
changed.

* modules/openstack_project/manifests/gerrit.pp: Build ref-updated from
a template and add an entry for the update_nametags.py file.

* modules/openstack_project/manifests/review.pp: Set the nametags
variable to a nullstring so that the update_nametags.py module won't be
added to review.openstack.org's change-merged hook.

* modules/openstack_project/manifests/review_dev.pp: Set the nametags
variable to reference openstack/release:nametags.yaml so that the
update_nametags.py module will be added to review-dev.openstack.org's
ref-updated hook.

* modules/openstack_project/templates/gerrit_ref-updated.erb:
Conditionally include a line for update_nametags.py only if a nametags
meta-branch:file pair has been specified for that host in puppet.

Change-Id: Ib1e6dd6cdfdf72b6ac1b9519c83f0d54678ffd1f