Reverse the close-pull logic.
We have a huge list of projects with all of them listing "close-pull" as an
option. Turns out we want to close pull requests for all of them - so instead,
let's only indicate 'has-pull-requests' if there is ever a project for whom
we do not want to close pull requests.

Change-Id: I4db31d324b7f260ea24f538ac5f5fc8e1feca75d