Do not reopen closed bugs due to bug mentions
Do not let update_bug.py reopen a FixCommitted or FixReleased bug just
because the bug number was mentioned in a commit message. In most cases
that mention is just a simple reference to a closed bug. In the rare
cases the committer actually wanted to reopen a closed bug, he should
rather have open a specific bug about it anyway.

Fixes bug 1078745

Change-Id: I513e6fc73d6bab02de21628e55a5d28189834632