Change diff command used by manage_projects.py.
manage_projects.py was running `git diff-index --quiet HEAD --`
previously to check if project.config had changed. This apparently
returns an exit code of 1 in some cases where a diff was not expected.
Switch to using `git diff --quiet HEAD` to check if any differences are
found. This appears to be more reliable in some manual testing.

Change-Id: I253423e41f80d71a5f2389bfc421e799f00f6fd9