Correct pypi upload form and copy artifact source
The file type form entry is 'filetype' not 'file_type'. Correct that in
pypi-upload.sh.

The artifact to copy is 'dist/project*.tar.gz' not 'project*.tar.gz'.
Correct that in pypi-jobs.yaml. Also cd into the dist dir that will be
created before attempting to copy the sdist out of that dir.

Change-Id: Ifa27e683c162f5acfc5136d0729be42af9b4ca41