Add gerritx project.
gerritx is the scripts that we use, mostly centered around projects.yaml
based activities, to manage additional functionality we need in a gerrit
installation. Since most of them are 100% general purpose, we broke them out
into a library so other people could use them.

A commit will follow this which will remove the scripts from the tree and
use pip-installed ones instead.

Change-Id: Iad460a9c08d09214d50adc04c777026af1ca48cb