Fix jeepyb install require on pypi.o.o.
The jeepyb install exec requires mysql::python but does not include that
class. This results in an error on hosts that do not include
mysql::python in some other manifest eg on pypi.openstack.org.

Include mysql::python in jeepyb's init manifest so that this require
works regardless of its inclusion with other manifests.

Change-Id: Iad1b1c723280df97ee5aa32512b36cfe1dc7034e