Use new etherpad-lite upstream.
etherpad-lite moved their upstream git repo to
https://github.com/ether/etherpad-lite.git. Update the vcsrepo source
for etherpad-lite when giving vcsrepo a specific version of
etherpad-lite. Do not update the source for the versionless vcsrepo
resource.

This allows us to leave etherpad.o.o alone while we install
etherpad-dev.o.o. Once etherpad-dev.o.o is working and we have tested an
upgrade to latest eplite and new node we can flip etherpad.o.o to the
new source.

Change-Id: I7cb9aaa5211f57c84698887500f519f92c514e91