Implement version filtering for oslo.
Oslo libraries want the following:

A tag to the repo will trigger tarball generation.

If the tag contains alpha\d*, beta\d* or rc\d*, the tarball will only
be uploaded to tarballs.openstack.org. This will allow those pre-releases
to be used in the other projects.

Otherwise, the tarball will be uploaded to tarballs.o.o AND PyPI. This
should allow us to use released versions of oslo libraries in pip-requires.

Change-Id: If99f048be36c249620368b19d8ca7504c698f29a