Add non-voting quantum devstack jobs.
Have two quantum devstack jobs, one that runs tempest,
and one that does not.  Both of them are non-voting.

Add the job that only runs exercises.sh to the set of projects in
the devstack-gate.

Also add the job that runs tempest to tempest, quantum, devstack,
and devstack-gate, to assist in continuing work to get tempest
working with quantum.

Change-Id: I9383ce2404c7fb31f43a14d7f2de721502aaa921