Copy tmp subunit log if file is present.
Copy .testrepository/tmp* with the scp plugin in python26 and python27
jobs if the file is present. This will give us some information about
which tests ran if testr hangs and is unable to finish within the
Jenkins timeout. If tests successfully run this file should not be
copied as testr deletes it before the plugin will attempt to copy it.

Change-Id: Id3d193094943e44bcf3083487f79a288d309d8c4