remove apt-get cmds from install_modules.sh
Updates install_modules so that we no longer try to install
puppet via apt-get here. This seems reasonable since we already
installed puppet 2.7 (which has 'module' support) in the
install_jenkins_slave.sh script.

The motivation for this change is to support multiple linux
distributions within openstack-infra and be able to run this
script on Red Hat linux distributions as well.

Change-Id: I1907624d1f934b832216261edf0acd87bfcc4bd2