Make an explicit full tempest job.
Instead of letting devstack-gate decide whether to run the full
suite or not. This makes it more clear what jobs are running,
and will provide better stats.

Also, remove some unneeded CINDER env vars.

This mostly attempts to keep the current behavior, however, it will
run only the smoke tests instead of the full test suite on the
postgres, cinder, and quantum jobs.

Change-Id: I234227b577a549b7b2243baeb072fe246637d7e3