Move ZUUL_BRANCH to params on periodic jobs.
The on-complete job needs the ZUUL_BRANCH passed to it in order
to accurately record statistics.  That's not possible if we
manually set that as an env variable in the shell script.  We
could use the setenv plugin, but it's just as easy (and has a
nice symmetry) to make it a parameter with a default value
(which is what the timed trigger will use).

Change-Id: Icd759bad15b722afe5fa539f050e630abf84ba9b
Reviewed-on: https://review.openstack.org/19754
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins