Add skeleton logstash module.
This new logstash module adds classes to install logstash agents and
indexers as well as redis and elasticsearch. The configuration for each
of these services is rudimentary but it shouldn't be difficult to make
to expand the configs and make them useful.

Also, add a logstash.openstack.org node that will have an agent,
indexer, redis, and elasticsearch installed on it. Eventually this
should get a logstash web front end as well.

Change-Id: I25b635f088f99d45cfaa70ed122c6433d3784937