Simplify and correct hiera install documentation
The documentation for hiera had the user add the repo and then remove it just
to install hiera, but that's now done in a previous step to use newer versions
of the puppet packages from the puppetlabs repos. The yaml file is also now
installed by puppet. Also collapsed all hiera documentation into the puppet
master section.

Change-Id: I00ab3cd12798295827ae25a623d246f7abf9a8df