make html less jumpy
clean up the html in the subunit2html style so that when you
open up sections the whole thing doesn't jump around on you
(i.e. column widths change, hot links drop around.)

Do this by shrinking the desc (don't repeat class name)

For table to 100% width, but shrink first column to 40%

Shrink all the number columns to 40px, and add an 8th colum to take
up the slack that happens when the results are shown.

Make the results div overflow-x: scroll, to ensure the page doesn't
get wider than the browser.

Change-Id: I9346d703778287fa7f1c72e2de077f225af070fb