Make the openstack_citest postgres user superuser.
Make the openstack_citest postgres user (added to Jenkins slaves'
postgres servers a postgres superuser). This will hopefully allow the
nova postgres tests to run.

Also, remove the database_grant that was previously being performed.
This was unnecessary as the postgresql::db['openstack_citest'] define
was taking care of the grant.

Change-Id: Ib43a4acd19f8460de7ddf9f593b0dcd73fb6467c