Setup logstash to use authenticated redis.
Require authentication for access to the logstash redis server and
update logstash agent and indexer configs to use authentication as well.

Also, make logstash agents talk to logstash.openstack.org instead of
localhost. This requires adding a list of iptables rules for the agents
that should be able to talk to redis on logstash.openstack.org's
external interface.

Change-Id: I98f86a7ca79f447efd367faef438c96e74015f07