Add baremetal DB migration test database.
Fixes bug 1130451

Add a second MySQL DB that the openstack_citest user has permissions to
to facilitate baremetal DB migration testing. This is necessary because
testr runs changes in parallel so each MySQL test needs individual DB
schemas. Long term we should come up with a better way of negotiating DB
access but in the short term this is quick and simple.

Change-Id: I4ee207bc3c2bf9d5d8c48409c2c611d17225994a