Move unattended_upgrades into site.pp.
Removes the 'include unattended_upgrades' from our
openstack_project modules and uses the site.pp to
include it only for 'Ubuntu' distros instead.

This should be a cleaner way to conditionally include this
only on Ubuntu.

For RHEL we want to do something similar by creating a
PackageKit-cron module and using that in site.pp.

Change-Id: I9209610582e23c86dccb5e70691bb598dad36950