Move tmpreaper into site.pp.
Removes the 'include tmpreaper' from our
openstack_project modules and uses the site.pp to
install it instead.

This should be a cleaner way to conditionally include this
only on Ubuntu pypi, template, and slave machines.

RHEL machines have a cron job setup to clean temp however we will
want to create a separate module with some finer grained controls
on this in a separate commit.

Change-Id: Ieea7a0884dbe511680c4af406e5ba064386975c2