change recheck to be age weighted
the number of rechecks against an item was artificially keeping things
at the top of the list well after the bug wasn't seen any more.
Change weighting to focus on age.

Also update the lp information on each bug deemed new enough so that
closed bugs are pushed further down the stack. In the future it would
be interesting to also weight based on importance.

Change-Id: Iaf76901725ddd89d3cea704c85da0d708c4b59b1