a more visually compact template for recheck
the recheck template was very expansive, which often meant that a
busy bug would take over the entire default window.

Instead of displaying things in <li>, make them lists on a single line.

Adjust whitespace on h3s so that they visually are attached to the
item they explain. Default h* spacing rules in html are actually
typically backwards, and require you to put a visual artifact above
them to imply the grouping you want.

Put a (closed) at the end of the title about bugs that are closed, to
give some indication about why they were shifted down the list.

It also deletes some errant trailing spaces, sorry about the unrelated
change there, but my emacs is configured to do that on every save.

Change-Id: I7869aa8a5b41f2f8826f61b8052933de659cb6cd