Add devstack-gate-secure.conf.
This file exists on the jenkins master but is not managed by puppet.
Put it on the new devstack-launch slave, under puppet control.

Rename the apikey hieradata variable that zuul uses to be more
generic since the same key is used by multiple projects.

Add the username portion of that to hiera as well to make the
hieradata file more self-documenting and easier to maintain, even
though it's not strictly secret.

The new values have been added to hiera.

Change-Id: I0af634d091577236a81feefddada63feacf3e2bc