Added configuration template to change from default db config to file-per-table config.
This is directed toward InnoDB. It appears there is an assumption that puppet-dashboard should be running in non-transaction safe mode, which is the default. This resolves the dependency for that assumption. The configuration template is copied from the mysql puppetlabs module and 'innodb_file_per_table' is appended to the bottom of the template.

Change-Id: I50dabea856f38fa6e8d4d3c589eb98f12a22457a