Fix mirror build jobs.
The job name matched the python27 functions regex, but as a post job,
it has no change/branch object.  Make the function safe for post jobs.

Configure them to also be triggerable by zuul (the timer still works
because all zuul variables are ignored).

Configure them to not run concurrently (there's only one matching
executor so it shouldn't be a problem, but be explicit) since they
share an output location.

Change-Id: I210da46bac1a8fa5ac9a931f07f189b3734b2ab0