Remove swift credentials from Jenkins slaves.
Glance "unit" tests previously connected to a remote swift object
store, but this test has not run in some time... known since the
credentials it was configured to use have been invalid for a while
and nobody noticed. These credentials are sensitive--or at least
would be if they were accurate--and thus should not be present on
Jenkins unit test slaves for security reasons.

Change-Id: Ic4341f7d53209af31c37a4207ab9eff00aa48879