Remove credentials for glance S3 testing.
Some glance "unit" tests connect to a remote S3 account, but this is
not in the spirit of proper unit testing and it's unclear whether
the tests are even run any longer. Also, it would be best not to
have credentials for remote services sitting on Jenkins unit test
slaves as they're accessible to any other tests and could be
trivially exposed.

Change-Id: I2cf76f9a77efc08598e803d3413bb719e84bfe6a