Install bundler on jenkins slaves
This commit adds bundler to the list of gems that need to
be installed on jenkins slave.

Bundler is the ruby equivalent of virtualenv, it allows tests to
install their own sets of gems into the local test directory
without attempting to install them as system gems. This is highly
recommended for any ruby unit tests that have their own sets of
gem dependencies (specifically, the puppet modules which will
be joining stackforge).

Change-Id: Ic443e6bf751cbf1604e4ab6b8910f45abb7eb8dd